refactor(cleanup): update devspace ide module name
