+++ pkg/controllers/robot_dev_suite/devspace_ide/check.go
@@ -1,4 +1,4 @@
-package robot_ide
+package devspace_ide
 
 import (
 	"context"
