+++ pkg/controllers/robot_dev_suite/devspace_ide/devspaceide_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package robot_ide
+package devspace_ide
 
 import (
 	"context"
